requirements.txt
# SmarterVote - All Python Dependencies
# Install with: pip install -r requirements.txt

# Pipeline Client (FastAPI backend)
fastapi==0.112.0
uvicorn[standard]==0.30.1
pydantic>=2.9.0
pydantic-settings==2.4.0
python-multipart==0.0.9
websockets>=12.0
httpx>=0.28.1,<1.0.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1

# AI client SDKs
openai>=1.84.0
anthropic>=0.40.0
google-genai>=1.69.0
google-auth>=2.48.1
requests>=2.28.1
tenacity>=8.2.3
google-cloud-storage==2.19.0

# Races API
# (uses fastapi/uvicorn/pydantic already above)

# Shared models
# pydantic already above

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
